Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float v;
-    scanf("%f",&v);
-    printf("You entered: %.4f",v);
+    double v;
+    scanf("%lf",&v);
+    printf("You entered: %g",v);
     return 0;
 }
